Support numbers in the randomInt function in java
The javascript implementation does, so we should here as well.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=289083119

+++ java/src/com/google/template/soy/basicfunctions/RandomIntFunction.java
@@ -61,7 +61,7 @@ public final class RandomIntFunction
   // lazy singleton pattern, allows other backends to avoid the work.
   private static final class Methods {
     static final Method RANDOM_INT_FN =
-        JavaValueFactory.createMethod(BasicFunctionsRuntime.class, "randomInt", long.class);
+        JavaValueFactory.createMethod(BasicFunctionsRuntime.class, "randomInt", double.class);
   }
 
   @Override
